Move update calculations to own file
This moves the updates calculations to its own file for
code and test.
